requirements.txt
django>=1.7,<1.8
djangorestframework>3.1,<3.2
django-rest-auth>0.3,<0.4
django-allauth>0.19,<0.20
Pillow>=2.7,<2.8
django-toolbelt>=0.0.1,<0.1
pytz
